requirements-flask.txt
flask==3.0.0
werkzeug==3.0.0
pypdf==3.17.2
google-generativeai==0.3.0
